+++ dflib/src/main/java/org/dflib/Udf0.java
@@ -15,6 +15,15 @@ package org.dflib;
 @FunctionalInterface
 public interface Udf0<R> {
 
+    /**
+     * A convenience "cast" method that allows to inline lambdas as a {@link Udf0}.
+     *
+     * @since 2.0.0
+     */
+    static <R> Udf0<R> of(Udf0<R> udf) {
+        return udf;
+    }
+
     Exp<R> call();
 
 }