+++ dflib/src/main/java/org/dflib/Udf1.java
@@ -11,6 +11,15 @@ import static org.dflib.Exp.$col;
 public interface Udf1<A, R> {
 
     /**
+     * A convenience "cast" method that allows to inline lambdas as a {@link Udf1}.
+     *
+     * @since 2.0.0
+     */
+    static <A, R> Udf1<A, R> of(Udf1<A, R> udf) {
+        return udf;
+    }
+
+    /**
      * Produces an expression based on the provided expression. This should be the preferred (though also most verbose)
      * way to resolve a UDF if the type of column is of significance. I.e. <code>call($int("a"))</code> will be faster
      * than <code>call("a")</code>, if the UDF converts the contents of "a" to ints internally.